+++ CHANGELOG.md
@@ -13,6 +13,13 @@
 - The event handler will no longer emit events with a position of `[0, 0, 0]` if no position could be determined.
   Instead, propagation is stopped on these events
 
+### 6.2.3
+
+- Updates `getFeaturesByCoordinate` in `wmsFeatureProvider` to consider the extent and to respect the `tilingSchema` in the bbox calculation
+- adds `htmlPositionFeatureTitle` to `wmsFeatureProvider` which is added as to the point feature created if (responseType === 'text/html').
+- Fixes an issue with COG layers, where there would be stripes in 3D.
+- Added option `sourceOptions` to COGLayer to forward options to the Openlayers source: https://openlayers.org/en/latest/apidoc/module-ol_source_GeoTIFF-GeoTIFFSource.html
+
 ### 6.2.2
 
 - Fixes an issue where the first panorama image was not shown in the correct resolution